requirements.txt
flask==3.0.0
google-genai>=1.57.0
pillow==10.4.0
python-dotenv==1.0.1
werkzeug==3.0.1
colorthief==0.2.1
